kcachegrind: Port to QGraphicsView
The cause of the slowdown before was the scaling down of the background
blending in the birds-eye view. This is fixed by using a QGradient there.

Improvements by taking advantage of the levelOfDetail in paint methods
of QGraphicsItems, e.g. scale line width accordingly and leave out
tiny names in the birds-eye view.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@701265 283d02a7-25f6-0310-bc7c-ecb5cbfe19da