kcachegrind: arrow thickness relates to call cost now
Take advantage of Qt4 painting features by adjusting
the arrow thickness in the call graph according to
the call cost (actually, thickness is logaritmic to cost).
The same is done in the panner view, which makes the
most used call chain quite obvious now.

Consequently, the call cost itself is not so interesting
any more as explicit annotation. More interesting is the
call count now. Cost percentage is still visible in the
icon. This somehow makes the call graph even less cleaner.

IMHO looks quite cool now ;-) I already wanted to do this
long ago, but this simply was not possible with Qt3, as it used the
line drawing in the X server, and thus the thickness was
not adjustable at all...


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@707069 283d02a7-25f6-0310-bc7c-ecb5cbfe19da