KCachegrind: fix for bug 161276
For detail level "normal", we want function nodes in the graph
to contain 2 lines of space for the symbol name, and the 3rd
line for cost display. Up to now (and also in KDE3 times), we
used hard coded scaling factors of the width/height numbers
given by dot. Thus, the number of lines was dependent on
the general KDE font *and* the default font used by dot for
space calculation. Of course, this can be different on every
system.
Now we calculate the needed scaling by looking at the general
KDE font, such that in "normal" detail level, it is guaranteed
that 3 lines fit into function nodes. In addition, we set a
maximal line number of the symbol space so that there is always
a line free for cost display.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@817266 283d02a7-25f6-0310-bc7c-ecb5cbfe19da