KCachegrind: reduce visualization view updates by update merging
The TraceViewItem class, which is the base class for all
visualization views in KCachegrind, now does update merging, ie.
a call to updateView() now does not directly result in a call
to doUpdate().

This e.g. cuts down the number of refreshes for the Flat profile
on startup of KCachegrind with a data file from 4 to 1, and also
gets rid of a short wrong display, as well as slow flicker.

Sometimes (e.g. when setting a new profile data model), we need
to skip this merging either completely for one update with
updateView(true), or for all updates of a TraceViewItem subclass
with setMergingUpdates(false).

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@1007541 283d02a7-25f6-0310-bc7c-ecb5cbfe19da