KCachegrind: Improve drawing performance of CallGraphView with Qt 4.6.0
Graphics items of the call graph are drawn in a simplified version in
the birds-eye view to keep drawing performance reasonable. For this, in
Qt 4.5.x, the <levelOfDetail> member of the QStyleOptionGraphicsItem
class, which is given to the painting function of graphics items, can be
used. This was obsoleted with Qt 4.6.0, and always gives 1 now, which
resulted in the same painting in the birds-eye view as in the main view.
The variable access is replaced by a function call now, taking the
transformation of the painter into account.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@1022106 283d02a7-25f6-0310-bc7c-ecb5cbfe19da