Add global option to hide C++ template parameters in symbols
This adds a toolbar button (+icon) and a menu item under "View".
When template parameters are hidden, you get the full symbol
by hovering over the active symbol name label. The later even
shows the structure of a C++ symbol.

Works both with KCachegrind and QCachegrind.

(done with the help of / partly by Michael Dengler)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@1078313 283d02a7-25f6-0310-bc7c-ecb5cbfe19da