Convert CoverageView from Q3ListView to QTreeWidget
Should have no behavioral changes.
For performance, it is important that the list is filled with
items at once, with sorting off.

This gets rid of the name argument of CoverageView constructor.
Instead, name settings is done Qt4-like with setObjectName().

(done with the help of / partly by Michael Dengler)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@1084378 283d02a7-25f6-0310-bc7c-ecb5cbfe19da