QCachegrind: start QFileDialog's in recent directory
This fixes the problem that File/Open and File/Add always
started up in the current working directory (where qcachegrind
was started), even if a file from another directory was opened.
We store the recently selected file, and use this as starting
point for the next QFileDialog.

The stored file name is not persistent over qcachegrind runs,
and the current directory will not be remembered when closing
a QFileDialog with 'Cancel'.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdesdk/kcachegrind@1095756 283d02a7-25f6-0310-bc7c-ecb5cbfe19da